requirements_minimal.txt
# FastAPI basics
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Image processing
pillow==10.1.0
numpy==1.24.3

# YOLO only (no torch installation needed - ultralytics handles it)
ultralytics==8.0.196

# For base64 decoding
pybase64==1.3.1